requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aiosqlite==0.20.0
async-timeout==4.0.3
attrs==24.2.0
docutils==0.21.2
et-xmlfile==1.1.0
frozenlist==1.4.1
greenlet==3.1.1
idna==3.10
lockfile==0.12.2
multidict==6.1.0
numpy==2.1.2
openpyxl==3.1.5
pandas==2.2.3
propcache==0.2.0
psycopg2-binary==2.9.9
python-daemon==3.0.1
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
SQLAlchemy==1.4.0
typing_extensions==4.12.2
tzdata==2024.2
yarl==1.15.2
